Updated orders

+++ src/main/java/Market/Repo/OrderRepository.java
@@ -3,6 +3,6 @@ package Market.Repo;
 import Market.Model.Orders;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Orders,Integer> {
+public interface OrderRepository extends JpaRepository<Orders,Long> {
 
 }
